Issue #268 - provide mvlc configuration tools
Continued testing of sis_vmusb_interface implementation ...this
found an error in the original -- in figuring out the number of words
actually written in the simulated block mode writes, constructs like

```
*written_nof_words++;
```

Really  just increment the pointer b/c ++ is higher prececence than *
Changed these to:

```
*written_nof_words+=1;
```

As I should have realized it should have been from the get-go.

Note this has also been done in the 7.1-dev branch of the repository.