Issue #268 - provide mvlc configuration tools
*  Protect the headers with MVLC_ include guards.
*  Add the headers we can to the build.

NOTE:  I don't think the SIS3316 class can be handed with a translator.
It's very interactive with the module.   It's going to have to wait
until we can actually include the support classes in the readout
software (making use of the initialization code he provided)...which
will be a big bit of work.

Could use a loadable .so to supply the interface support.