Issue #288 -  Use A24 AMs for CAEN 32 channel modules for 0x00xx bases
I don't think this will make a difference but theoretically, the
geo register is only writable for modules that don't have PAUX and
this can be tested for with Status1's AMNESIA bit.

Why I don't think it matters.  Initialization is done via individual
operations, not lists and statuses of each of the write operations is
not checked so what would happen normally is that the

```c++
controller.vmeWrite16(base+GEO, initamod, geo);
```

Would fail - silently and inidialization would continue.
This _does_ matter for the mvlc translator which needs to be told,
for these modules that there's a PAUX and therefore don't write this
register -- that's the failure I saw in mvlcreadout