MVLC Initialization of V785,792, 775 failed:
The failure was due to writing the GEO register when there's
a PAUX connector.  What I did:

*   Add a -paux boolean option that is ```true``` by default.
*   Only write the geo register with the -geo parameter if that's
    ```false```

note that at FRIB most backplanes have a PAUX connector, hence the
default.  If you don't have a PAUX, then set this option false and
Geo will be written to identify the module.