Issue #294 - Add VMSUB workarounds
Move the definition of the conditionals
to configure.ac so COutputThread can also see it.  Three new items:

--disable-vmusb-prompt-busy
--disable-vmusb-end-event
--disable-vmusb-idle-busy

so the workarounds are on by default.  Note that it is easy enough to
change these to --enable's so that they are off by default.  Simply
change the help text and the if not present action to xxx="no"