Shove headers into @prefix/include/sbsmodules
this is needed to prevent collisions with e.g. CCAENChain while buiding
mvlc_generate:

That's because the default header directories in compilation include
@prefix@/include and we want CCAENChain included from
@top_srcdir@/usb/vmusb/daqconfig

Since few, if any, are using the sbsReadout with SBSModules I don't
think this causes problems.  If they do, there are two fixes

 #include <sbsmodules/whathaveyou.h>

or
  -I$DAQINC/sbsmodules on the compilation command and no source
mods are needed.